// internship-frontend/src/services/openapi/models/User.ts
export type User = {
    id: number;
    firstName: string;
    middleName: string;
    lastName: string;
    commonName: string;
    homeAddress: string;
    gender: User.gender;
    dateOfBirth: string;
    nationality: User.nationality;
    state: string;
    city: string;
    county: string;
    zip: string;
    isActive: boolean;
    isSoftDeleted: boolean;
    primaryEmailAddress: string;
    backupEmailAddress: string;
    phone: {
        mobile?: Array<string>,
        office?: Array<string>,
        home?: Array<string>,
    };
    isPrimaryEmailAddressVerified: boolean;
    isBackupEmailAddressVerified: boolean;
    passwordHash: string;
    isPasswordChangeRequired: boolean;
    resetPasswordToken: string;
    resetPasswordExpiration: string;
    primaryEmailVerificationToken: string;
    backupEmailVerificationToken: string;
    emailVerificationTokenExpiration: string;
    otpEnabled: boolean;
    otpSecret: string;
    refreshTokenHash: string;
    employeeId: number;
    employee: Employee;
    userProfile: UserProfile;
    roles: Array<Role>;
}

export namespace User {

    export enum gender {
        MALE = 'male',
        FEMALE = 'female',
    }
// ...
}